move types definitions outside of cpp
