+++ src/update-client.cc
@@ -34,17 +34,7 @@ using std::regex_search;
 #include "update-parameters.hpp"
 #include "logger/log.h"
 
-/*##############################################
- *#
- *# Update exceptions
- *#
- *############################################*/
-
-class update_exception_blocked : public std::exception
-{};
-
-class update_exception_failed : public std::exception
-{};
+#include "update-client-internal.hpp"
 
 /*##############################################
  *#
@@ -73,183 +63,6 @@ namespace {
 	}
 
 }
-
-/*##############################################
- *#
- *# Data definitions
- *#
- *############################################*/
-struct update_client {
-	template <class Body, bool IncludeVersion>
-	struct http_request;
-
-	struct file;
-	struct pid;
-
-	using manifest_body = http::basic_dynamic_body<beast::flat_buffer>;
-
-	template <class Body>
-	using manifest_request = http_request<Body, false>;
-
-	template <class Body>
-	using file_request = http_request<Body, true>;
-
-	template <class Type>
-	using executor_work_guard = asio::executor_work_guard<Type>;
-	using io_context = asio::io_context;
-	using work_guard_type = executor_work_guard<io_context::executor_type>;
-	using resolver_type = tcp::resolver;
-	using manifest_map = std::unordered_map<std::string, std::string>;
-
-	update_client() = delete;
-	update_client(const update_client&) = delete;
-	update_client(update_client&&) = delete;
-	update_client &operator=(update_client&&) = delete;
-	update_client &operator=(const update_client&) = delete;
-
-	explicit update_client(struct update_parameters *params);
-	~update_client();
-
-	void set_client_events(client_callbacks *cbs) { client_events = cbs; }
-
-	void set_downloader_events(downloader_callbacks *cbs) { downloader_events = cbs; }
-
-	void set_updater_events(updater_callbacks *cbs) { updater_events = cbs; }
-
-	void set_pid_events(pid_callbacks *cbs) { pid_events = cbs; }
-
-	void do_stuff();
-	void flush();
-
-	/* These two are explicitly initialized in constructor */
-	io_context io_ctx;
-	update_parameters       *params;
-
-	work_guard_type         *work{ nullptr };
-	fs::path                 new_files_dir;
-	client_callbacks        *client_events{ nullptr };
-	downloader_callbacks    *downloader_events{ nullptr };
-	updater_callbacks       *updater_events{ nullptr };
-	pid_callbacks           *pid_events{ nullptr };
-
-	int                      active_workers{ 0 };
-	std::atomic_size_t       active_pids{ 0 };
-	manifest_map             manifest;
-	std::mutex               manifest_mutex;
-	manifest_map::const_iterator manifest_iterator;
-
-	resolver_type            resolver;
-	resolver_type::results_type endpoints;
-	ssl::context ssl_context{ ssl::context::method::sslv23_client };
-
-	std::vector<std::thread> thread_pool;
-
-private:
-	inline void handle_error(const boost::system::error_code &error, const char* str);
-	void handle_file_download_error(file_request<http::dynamic_body> *request_ctx, const boost::system::error_code &error, const char* str);
-
-	void start_downloader();
-	//manifest 
-	void handle_resolve(const boost::system::error_code &error, tcp::resolver::results_type results);
-
-	void handle_manifest_connect(const boost::system::error_code &error, const tcp::endpoint &ep, manifest_request<manifest_body> *request_ctx);
-
-	void handle_manifest_handshake(const boost::system::error_code &error, manifest_request<manifest_body> *request_ctx);
-
-	void handle_manifest_request(boost::system::error_code &error, size_t bytes, manifest_request<manifest_body> *request_ctx);
-
-	void handle_manifest_response(boost::system::error_code &ec, size_t bytes, manifest_request<manifest_body> *request_ctx);
-
-	void handle_manifest_results();
-
-	void clean_manifest();
-	void check_file(fs::path & check_path);
-
-	//files
-	void handle_manifest_entry(file_request<http::dynamic_body> *request_ctx);
-
-	void handle_file_connect(const boost::system::error_code &error, const tcp::endpoint &ep, file_request<http::dynamic_body> *request_ctx);
-
-	void handle_file_handshake(const boost::system::error_code& error, file_request<http::dynamic_body> *request_ctx);
-
-	void handle_file_request(boost::system::error_code &error, size_t bytes, file_request<http::dynamic_body> *request_ctx);
-
-	void handle_file_response_header(boost::system::error_code &error, size_t bytes, file_request<http::dynamic_body> *request_ctx);
-
-	void handle_file_response_body(boost::system::error_code &error, size_t bytes_read, file_request<http::dynamic_body> *request_ctx, update_client::file *file_ctx);
-
-	void handle_file_result(update_client::file *file_ctx, int index);
-
-	//update 
-	void handle_pids();
-
-	void handle_pid(const boost::system::error_code& error, update_client::pid* contexts);
-
-	void start_file_update();
-
-	void create_work();
-	void reset_work();
-};
-
-template <class Body, bool IncludeVersion>
-struct update_client::http_request
-{
-	http_request(update_client *client_ctx, const std::string &target, const int id);
-	~http_request();
-
-	size_t download_accum{ 0 };
-	size_t content_length{ 0 };
-	int worker_id;
-	update_client *client_ctx;
-	std::string target;
-
-
-	/* We used to support http and then I realized
-	 * I was spending a lot of time supporting both.
-	 * Our use case doesn't use it and boost doesn't
-	 * allow any convenience to allow using them
-	 * interchangeably. Really, my suggestion is that
-	 * you should be using ssl regardless anyways. */
-	ssl::stream<tcp::socket> ssl_socket;
-	
-	/* We need way to detect stuck connection
-	*  For that we use boost deadline timer what can limit 
-	*  time for each step of file downloader connection
-	*  Also limit recieve buffer so timer limit too slow fill of the buffer  
-	*/
-	boost::asio::deadline_timer deadline;
-	int deadline_default_timeout = 5;
-	bool deadline_reached = false;
-	int retries = 0;
-	http::request<http::empty_body> request;
-
-	beast::multi_buffer response_buf;
-	http::response_parser<Body> response_parser;
-
-	void check_deadline_callback(const boost::system::error_code& error);
-};
-
-struct FileUpdater
-{
-	FileUpdater() = delete;
-	FileUpdater(FileUpdater&&) = delete;
-	FileUpdater(const FileUpdater&) = delete;
-	FileUpdater &operator=(const FileUpdater&) = delete;
-	FileUpdater &operator=(FileUpdater&&) = delete;
-
-	explicit FileUpdater(update_client *client_ctx);
-	~FileUpdater();
-
-	void update();
-	void update_entry(update_client::manifest_map::iterator  &iter, boost::filesystem::path & new_files_dir);
-	void revert();
-	bool reset_rights(const fs::path& path);
-
-private:
-	update_client *m_client_ctx;
-	fs::path m_old_files_dir;
-	fs::path m_app_dir;
-};
 
 /*##############################################
  *#
@@ -364,7 +177,7 @@ void FileUpdater::revert()
 
 // limit response_buf buffer size so together with deadline timeout 
 // it will create low speed limit for download 
-// 512kb in 10 seconds is somewhere of old modems 
+// 4kb in 4 seconds is somewhere of old modems 
 // and be recognized as stuck connection
 
 template <class Body, bool IncludeVersion>
@@ -373,24 +186,17 @@ update_client::http_request<Body, IncludeVersion>::http_request(update_client *c
 	client_ctx(client_ctx),
 	target(target),
 	ssl_socket(client_ctx->io_ctx, client_ctx->ssl_context),
-	deadline(client_ctx->io_ctx),	
-	response_buf(4*1024) // see reasons above ^
+	deadline(client_ctx->io_ctx),
+	response_buf(4 * 1024) // see reasons above ^
 {
 	std::string full_target;
 
 	if (IncludeVersion)
 	{
-		full_target = fmt::format("{}/{}/{}",
-			client_ctx->params->host.path,
-			client_ctx->params->version,
-			target
-		);
+		full_target = fmt::format("{}/{}/{}", client_ctx->params->host.path, client_ctx->params->version, target);
 	}
 	else {
-		full_target = fmt::format("{}/{}",
-			client_ctx->params->host.path,
-			target
-		);
+		full_target = fmt::format("{}/{}", client_ctx->params->host.path, target);
 	}
 
 	request = { http::verb::get, full_target, 11 };
@@ -488,11 +294,7 @@ void update_client::handle_pids()
 		start_file_update();
 	}
 
-	for (
-		auto iter = params->pids.begin();
-		iter != params->pids.end();
-		++iter
-		)
+	for (auto iter = params->pids.begin(); iter != params->pids.end(); ++iter)
 	{
 		HANDLE hProcess = OpenProcess(PROCESS_ALL_ACCESS, FALSE, (DWORD)*iter);
 
@@ -541,37 +343,26 @@ void update_client::handle_file_download_error(file_request<http::dynamic_body> 
 /* FIXME Make the other handlers part of the client class. */
 void update_client::handle_resolve(const boost::system::error_code &error, tcp::resolver::results_type results) {
 	if (error) {
-		client_events->error(
-			"Failed to connect to update server."
-		);
+		client_events->error("Failed to connect to update server.");
 		return;
 	}
 
 	endpoints = results;
 
 	/* TODO I should make hash type configurable. */
-	std::string manifest_target{
-		fmt::format("{}.sha256", params->version)
-	};
+	std::string manifest_target{ fmt::format("{}.sha256", params->version) };
 
 	/* This seems like duplicate code but notice that
 	 * request_ctx becomes two different types depending
 	 * on the path. We only need to do this once, the rest
 	 * is automatically deduced. */
-	auto *request_ctx = new manifest_request<manifest_body>(
-		this, manifest_target, 0
-		);
+	auto *request_ctx = new manifest_request<manifest_body>(this, manifest_target, 0);
 
-	auto connect_handler =
-		[this, request_ctx](auto i, auto e) {
+	auto connect_handler = [this, request_ctx](auto i, auto e) {
 		this->handle_manifest_connect(i, e, request_ctx);
 	};
 
-	asio::async_connect(
-		request_ctx->ssl_socket.lowest_layer(),
-		results,
-		connect_handler
-	);
+	asio::async_connect(request_ctx->ssl_socket.lowest_layer(), results, connect_handler);
 }
 
 update_client::update_client(struct update_parameters *params)
@@ -590,9 +381,11 @@ update_client::update_client(struct update_parameters *params)
 
 	work = new work_guard_type(asio::make_work_guard(io_ctx));
 
-	for (unsigned i = 0; i < num_workers; ++i) {
+	for (unsigned i = 0; i < num_workers; ++i)
+	{
 		thread_pool.emplace_back(
-			std::thread([=]() {
+			std::thread([=]()
+		{
 			io_ctx.run();
 		})
 		);
@@ -621,14 +414,16 @@ void update_client::reset_work()
 
 void update_client::flush()
 {
-	for (std::thread &thrd : thread_pool) {
+	for (std::thread &thrd : thread_pool)
+	{
 		thrd.join();
 	}
 }
 
 void update_client::do_stuff()
 {
-	auto cb = [=](auto e, auto i) {
+	auto cb = [=](auto e, auto i)
+	{
 		this->handle_resolve(e, i);
 	};
 
@@ -657,10 +452,7 @@ static std::string calculate_checksum(fs::path &path)
 
 	char useless_buffer[4096];
 
-	checksum_chain.push(
-		boost::reference_wrapper<sha256_filter>(checksum_filter),
-		4096
-	);
+	checksum_chain.push(boost::reference_wrapper<sha256_filter>(checksum_filter), 4096);
 
 	checksum_chain.push(bio::file_descriptor_source(path), 4096);
 
@@ -674,11 +466,9 @@ static std::string calculate_checksum(fs::path &path)
 	hex_digest.reserve(64);
 
 	/* TODO 32 is hardcoded here for the size of an SHA-256 digest buffer */
-	for (int i = 0; i < 32; ++i) {
-		fmt::format_to(
-			std::back_inserter(hex_digest),
-			"{:02x}", checksum_filter.digest[i]
-		);
+	for (int i = 0; i < 32; ++i)
+	{
+		fmt::format_to(std::back_inserter(hex_digest), "{:02x}", checksum_filter.digest[i]);
 	}
 
 	return hex_digest;
@@ -697,9 +487,7 @@ void update_client::clean_manifest()
 	{
 		fs::path entry = app_dir_iter->path();
 
-		fs::path key_path(
-			fs::relative(entry, this->params->app_dir)
-		);
+		fs::path key_path(fs::relative(entry, this->params->app_dir));
 
 		std::string key = key_path.make_preferred().string();
 
@@ -716,7 +504,8 @@ void update_client::clean_manifest()
 		/* If the checksum is the same as in the manifest,
 		 * remove it from the manifest entirely, there's no need
 		 * to download it (as it's already the same). */
-		if (checksum.compare(manifest_iter->second) == 0) {
+		if (checksum.compare(manifest_iter->second) == 0)
+		{
 			this->manifest.erase(manifest_iter);
 			continue;
 		}
@@ -888,10 +677,7 @@ void update_client::handle_manifest_response(boost::system::error_code &ec, size
 
 	if (ec)
 	{
-		std::string msg = fmt::format(
-			"Failed manifest response ({})",
-			safe_request_ctx->target
-		);
+		std::string msg = fmt::format("Failed manifest response ({})", safe_request_ctx->target);
 
 		handle_error(ec, msg.c_str());
 		return;
@@ -919,10 +705,7 @@ void update_client::handle_manifest_request(boost::system::error_code &error, si
 {
 	if (error)
 	{
-		std::string msg = fmt::format(
-			"Failed manifest request ({}): {}",
-			request_ctx->target, error.message().c_str()
-		);
+		std::string msg = fmt::format("Failed manifest request ({}): {}", request_ctx->target, error.message().c_str());
 
 		handle_error(error, msg.c_str());
 		return;
@@ -946,10 +729,7 @@ void update_client::handle_manifest_handshake(const boost::system::error_code& e
 {
 	if (error)
 	{
-		std::string msg = fmt::format(
-			"Failed manifest handshake ({}): {}",
-			request_ctx->target, error.message().c_str()
-		);
+		std::string msg = fmt::format("Failed manifest handshake ({}): {}", request_ctx->target, error.message().c_str());
 
 		handle_error(error, msg.c_str());
 	}
@@ -1056,10 +836,7 @@ void update_client::handle_file_result(update_client::file *file_ctx, int index)
 		 * an SHA-256 digest buffer */
 		for (int i = 0; i < 32; ++i)
 		{
-			fmt::format_to(
-				std::back_inserter(hex_digest),
-				"{:02x}", filter.digest[i]
-			);
+			fmt::format_to(std::back_inserter(hex_digest), "{:02x}", filter.digest[i]);
 		}
 	}
 	catch (...)
@@ -1239,10 +1016,7 @@ void update_client::handle_file_handshake(const boost::system::error_code& error
 {
 	if (error)
 	{
-		std::string msg = fmt::format(
-			"Failed manifest handshake ({}): {}",
-			request_ctx->target, error.message().c_str()
-		);
+		std::string msg = fmt::format("Failed manifest handshake ({}): {}", request_ctx->target, error.message().c_str());
 
 		handle_file_download_error(request_ctx, error, msg.c_str());
 		return;
